Move zcashTest transaction tests to forks/
Reduce diff to upstream
