+++ test/transaction.js
@@ -4,7 +4,6 @@ var assert = require('assert')
 var bscript = require('../src/script')
 var networks = require('../src/networks')
 var fixtures = require('./fixtures/transaction')
-var fixturesZec = require('./fixtures/forks/zcashTest/transaction')
 var Transaction = require('../src/transaction')
 
 describe('Transaction', function () {
@@ -86,56 +85,6 @@ describe('Transaction', function () {
       var tx = Transaction.fromHex(txHex)
       assert.equal(-1, tx.version)
       assert.equal(0xffffffff, tx.locktime)
-    })
-  })
-
-  describe('fromBuffer/fromHex for Zcash', function () {
-    fixturesZec.valid.forEach(function (testData) {
-      it('imports ' + testData.description, function () {
-        const tx = Transaction.fromHex(testData.hex, networks.zcashTest)
-        assert.equal(tx.version, testData.version)
-        assert.equal(tx.versionGroupId, parseInt(testData.versionGroupId, 16))
-        assert.equal(tx.overwintered, testData.overwintered)
-        assert.equal(tx.locktime, testData.locktime)
-        assert.equal(tx.expiryHeight, testData.expiryHeight)
-        assert.equal(tx.ins.length, testData.insLength)
-        assert.equal(tx.outs.length, testData.outsLength)
-        assert.equal(tx.joinsplits.length, testData.joinsplitsLength)
-        assert.equal(tx.joinsplitPubkey.length, testData.joinsplitPubkeyLength)
-        assert.equal(tx.joinsplitSig.length, testData.joinsplitSigLength)
-
-        if (testData.valueBalance) {
-          assert.equal(tx.valueBalance, testData.valueBalance)
-        }
-        if (testData.nShieldedSpend > 0) {
-          for (var i = 0; i < testData.nShieldedSpend; ++i) {
-            assert.equal(tx.vShieldedSpend[i].cv.toString('hex'), testData.vShieldedSpend[i].cv)
-            assert.equal(tx.vShieldedSpend[i].anchor.toString('hex'), testData.vShieldedSpend[i].anchor)
-            assert.equal(tx.vShieldedSpend[i].nullifier.toString('hex'), testData.vShieldedSpend[i].nullifier)
-            assert.equal(tx.vShieldedSpend[i].rk.toString('hex'), testData.vShieldedSpend[i].rk)
-            assert.equal(tx.vShieldedSpend[i].zkproof.sA.toString('hex') +
-              tx.vShieldedSpend[i].zkproof.sB.toString('hex') +
-              tx.vShieldedSpend[i].zkproof.sC.toString('hex'), testData.vShieldedSpend[i].zkproof)
-            assert.equal(tx.vShieldedSpend[i].spendAuthSig.toString('hex'), testData.vShieldedSpend[i].spendAuthSig)
-          }
-        }
-      })
-    })
-
-    fixturesZec.valid.forEach(function (testData) {
-      it('exports ' + testData.description, function () {
-        const tx = Transaction.fromHex(testData.hex, networks.zcashTest)
-        const hexTx = tx.toHex()
-        assert.equal(testData.hex, hexTx)
-      })
-    })
-
-    fixturesZec.valid.forEach(function (testData) {
-      it('clone ' + testData.description, function () {
-        const tx = Transaction.fromHex(testData.hex, networks.zcashTest)
-        const clonedTx = tx.clone()
-        assert.equal(clonedTx.toHex(), testData.hex)
-      })
     })
   })
 
@@ -378,35 +327,6 @@ describe('Transaction', function () {
       assert.throws(function () {
         (new Transaction()).setWitness(0, 'foobar')
       }, /Expected property "1" of type \[Buffer], got String "foobar"/)
-    })
-  })
-
-  describe('hashForZcashSignature', function () {
-    fixturesZec.hashForZcashSignature.valid.forEach(function (testData) {
-      it('should return ' + testData.hash + ' for ' + testData.description, function () {
-        var network = networks.zcash
-        network.consensusBranchId[testData.version] = parseInt(testData.branchId, 16)
-        var tx = Transaction.fromHex(testData.txHex, network)
-        var script = Buffer.from(testData.script, 'hex')
-        var hash = Buffer.from(testData.hash, 'hex')
-        hash.reverse()
-        hash = hash.toString('hex')
-
-        assert.strictEqual(
-          tx.hashForZcashSignature(testData.inIndex, script, testData.value, testData.type).toString('hex'),
-          hash)
-      })
-    })
-
-    fixturesZec.hashForZcashSignature.invalid.forEach(function (testData) {
-      it('should throw on ' + testData.description, function () {
-        var tx = Transaction.fromHex(testData.txHex, networks.zcashTest)
-        var script = Buffer.from(testData.script, 'hex')
-
-        assert.throws(function () {
-          tx.hashForZcashSignature(testData.inIndex, script, testData.value, testData.type)
-        }, new RegExp(testData.exception))
-      })
     })
   })
 })
